pyproject.toml
[project]
name = "aura-v2"
version = "2.0.0"
description = "AURA Enterprise v2 – real-time multi-object tracking"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "typer>=0.9",
  "fastapi>=0.116",
  "uvicorn>=0.23",
  "pydantic>=2.5",
  "pyyaml>=6.0",
  "numpy>=1.26",
  "filterpy>=1.4.5",
  "scipy>=1.11",
  "scikit-learn>=1.3",
  "grpcio>=1.59",
  "websockets>=12.0",
  "opentelemetry-api>=1.20",
  "opentelemetry-sdk>=1.20",
  "prometheus-client>=0.22",
  "structlog>=25",
  "aioredis>=2.0",
  "aiokafka>=0.12",
  "motor>=3.7",
  "dependency-injector>=4.41.0",
  # Runtime pin so Starlette TestClient and dev source-pump stay compatible
  "httpx>=0.27,<0.29",
  "httpcore>=1.0.2",
]

[project.optional-dependencies]
tracking = [
  "ultralytics>=8.0",
  "boxmot>=11,<13; python_version < '3.13'",
  "PyQt5; sys_platform == 'win32'",
  "pyqt5-qt5; sys_platform != 'win32'",
]

observability = [
  "opentelemetry-api>=1.20",
  "opentelemetry-sdk>=1.20",
  "prometheus-client>=0.22",
  "structlog>=25",
]

messaging = [
  "aiokafka>=0.12",
  "aioredis>=2.0",
  "motor>=3.7",
]

perf = []

test = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23,<0.25",
  "anyio>=3.7,<5",
]

# Linux-only GPU extras; will never resolve on macOS
gpu = [
  "cupy-cuda12x>=13; platform_system == 'Linux'",
  "torch>=2.2; platform_system == 'Linux'",
  "tensorrt>=8.6; platform_system == 'Linux'",
  "tensorrt-cu13==10.13.3.9; platform_system == 'Linux' and platform_machine == 'x86_64'",
]

[project.scripts]
aura-cli = "aura_v2.main:app_cli"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["aura_v2*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
]
